+++ package.json
@@ -19,6 +19,7 @@
     "konva": "^8.4.3",
     "lz-string": "^1.5.0",
     "node-sass": "^8.0.0",
+    "npyjs": "^0.4.0",
     "onnxruntime-common": "^1.14.0",
     "onnxruntime-web": "^1.14.0",
     "process": "^0.11.10",
