+++ src/components/helpers/photos.tsx
@@ -1,170 +1,17 @@
-const photos = [
-    {
-      src: "/assets/gallery/18135.jpg",
-      width: 2000,
-      height: 1538,
-    },
-    {
-      src: "/assets/gallery/AdobeStock_94274587_welsh_corgi_pembroke_CD.jpg",
-      width: 2048,
-      height: 1365,
-    },
-    {
-      src: "/assets/gallery/farmhouse_in_provence_1970.17.34.jpg",
-      width: 3000,
-      height: 2250,
-    },
-    {
-      src: "/assets/gallery/GettyImages-587358024.jpg",
-      width: 2048,
-      height: 1364,
-    },
-    {
-      src: "/assets/gallery/GettyImages-673968470.jpg",
-      width: 2068,
-      height: 1450,
-    },
-    {
-      src: "/assets/gallery/GettyImages-1191014275.jpg",
-      width: 2500,
-      height: 1661,
-    },
-    {
-      src: "/assets/gallery/GettyImages-1207721867.jpg",
-      width: 2048,
-      height: 1367,
-    },
-    {
-      src: "/assets/gallery/GettyImages-1207911110.jpg",
-      width: 2048,
-      height: 1349,
-    },
-    {
-      src: "/assets/gallery/IMG_0378.jpg",
-      width: 3000,
-      height: 2103,
-    },
-    {
-      src: "/assets/gallery/shutterstock_466589609.jpg",
-      width: 2769,
-      height: 2005,
-    },
-    {
-      src: "/assets/gallery/AdobeStock_402316.jpg",
-      width: 3295,
-      height: 1846,
-    },
-    {
-      src: "/assets/gallery/AdobeStock_57083879.jpg",
-      width: 3000,
-      height: 2000,
-    },
-    {
-      src: "/assets/gallery/AdobeStock_135146289.jpg",
-      width: 3000,
-      height: 2000,
-    },
-    {
-      src: "/assets/gallery/AdobeStock_323574125.jpg",
-      width: 2240,
-      height: 3144,
-    },
-    {
-      src: "/assets/gallery/FBB_Apartment_BE16_DSC03308.jpg",
-      width: 3000,
-      height: 2189,
-    },
-    {
-      src: "/assets/gallery/GettyImages-104691018.jpg",
-      width: 3000,
-      height: 1999,
-    },
-    {
-      src: "/assets/gallery/dogs-with-stick.jpg",
-      width: 1072,
-      height: 603,
-    },
-    {
-      src: "/assets/gallery/GettyImages-182510030.jpg",
-      width: 3000,
-      height: 2188,
-    },
-    {
-      src: "/assets/gallery/GettyImages-487967520.jpg",
-      width: 3000,
-      height: 2000,
-    },
-    {
-      src: "/assets/gallery/GettyImages-523306132.jpg",
-      width: 3000,
-      height: 2357,
-    },
-    {
-      src: "/assets/gallery/GettyImages-540782045.jpg",
-      width: 3000,
-      height: 2127,
-    },
-    {
-      src: "/assets/gallery/GettyImages-595762959.jpg",
-      width: 3945,
-      height: 5523,
-    },
-    {
-      src: "/assets/gallery/GettyImages-636036276.jpg",
-      width: 3000,
-      height: 2248,
-    },
-    {
-      src: "/assets/gallery/GettyImages-901017386.jpg",
-      width: 3000,
-      height: 1948,
-    },
-    {
-      src: "/assets/gallery/GettyImages-941086978.jpg",
-      width: 2302,
-      height: 2302,
-    },
-    {
-      src: "/assets/gallery/GettyImages-1058471994.jpg",
-      width: 3000,
-      height: 2066,
-    },
-    {
-      src: "/assets/gallery/GettyImages-1124460942.jpg",
-      width: 3000,
-      height: 2000,
-    },
-    {
-      src: "/assets/gallery/GettyImages-1207391022.jpg",
-      width: 3000,
-      height: 2099,
-    },
-    {
-      src: "/assets/gallery/GettyImages-1215928137.jpg",
-      width: 3000,
-      height: 2621,
-    },
-    {
-      src: "/assets/gallery/GettyImages-1216603285.jpg",
-      width: 3000,
-      height: 2027,
-    },
-    {
-      src: "/assets/gallery/iStock-973619282.jpg",
-      width: 3000,
-      height: 2038,
-    },
-    {
-      src: "/assets/gallery/shutterstock_1385445275.jpg",
-      width: 2737,
-      height: 2083,
-    },
-    {
-      src: "/assets/gallery/shutterstock_1723139395.jpg",
-      width: 3000,
-      height: 2155,
-    },
-  ];
-  
-  export default photos;
-  
+import { API_LIST_ALL } from "../../enviroments";
+
+
+const getPhotos = (): any[] => {
+  const xhr = new XMLHttpRequest();
+  xhr.open("GET", API_LIST_ALL, false);
+  xhr.send();
+  const json = JSON.parse(xhr.response);
+  // console.log("GOT LIST OF FILES");
+  // console.log(json);
+  // console.log("DONE LIST OF FILES");
+  return json["files"];
+}
+
+const photos = getPhotos();
+
+export default photos;
